ai_law_chatbot/backend/pyproject.toml
[project]
name = "ai_law_chatbot"
version = "0.1.0"
description = "My AI law chatbot"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "doanlan96", email = "ducpc0602@gmail.com" }]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "orjson>=3.10.0",
    "logfire[fastapi,asyncpg,redis,httpx]>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",
    "greenlet>=3.0.0",
    "pyjwt>=2.9.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.12",
    "authlib>=1.3.0",
    "httpx>=0.27.0",
    "redis>=5.2.0",
    "fastapi-cache2>=0.2.2",
    "slowapi>=0.1.9",
    "fastapi-pagination>=0.12.31",
    "sentry-sdk[fastapi]>=2.18.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "sqladmin>=0.19.0",
    "itsdangerous>=2.2.0",
    "boto3>=1.35.0",
    "pydantic-ai-slim[openrouter]>=0.0.39",
    "httpx>=0.27.0",
    "click>=8.1.0",
    "tabulate>=0.9.0",
    "pgvector>=0.4.2",
    "google-generativeai>=0.8.6",
    "pypdf>=6.9.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "anyio[trio]>=4.0.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
ai_law_chatbot = "cli.commands:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "cli"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (needed for FastAPI Depends)
]

[tool.ruff.lint.isort]
known-first-party = ["app", "cli"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.fastapi-fullstack]
generator_version = "0.2.1"
generated_at = "2026-03-09T14:41:19.208917+00:00"
